+++ media-sample/src/androidTest/java/com/google/android/horologist/mediasample/runner/FakeConfigModule.kt
@@ -23,11 +23,11 @@ import android.os.StrictMode
 import com.google.android.horologist.audio.SystemAudioRepository
 import com.google.android.horologist.media3.audio.AudioOutputSelector
 import com.google.android.horologist.media3.audio.BluetoothSettingsOutputSelector
-import com.google.android.horologist.mediasample.AppConfig
 import com.google.android.horologist.mediasample.di.CacheDir
 import com.google.android.horologist.mediasample.di.ConfigModule
 import com.google.android.horologist.mediasample.di.IsEmulator
-import com.google.android.horologist.mediasample.util.resetAfter
+import com.google.android.horologist.mediasample.ui.AppConfig
+import com.google.android.horologist.mediasample.ui.util.resetAfter
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.android.qualifiers.ApplicationContext
