+++ media-sample/src/main/java/com/google/android/horologist/mediasample/data/log/Logging.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.mediasample.system
+package com.google.android.horologist.mediasample.data.log
 
 import android.content.res.Resources
 import android.util.Log
