+++ media-sample/src/main/java/com/google/android/horologist/mediasample/data/service/complication/DataUpdates.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.mediasample.complication
+package com.google.android.horologist.mediasample.data.service.complication
 
 import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
