+++ media-sample/src/main/java/com/google/android/horologist/mediasample/data/service/complication/MediaStatusComplicationService.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Android Open Source Project
+ * Copyright 2022 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.mediasample.complication
+package com.google.android.horologist.mediasample.data.service.complication
 
 import android.graphics.drawable.Icon
 import androidx.media3.common.MediaItem
