+++ media-sample/src/main/java/com/google/android/horologist/mediasample/di/ConfigModule.kt
@@ -23,8 +23,8 @@ import android.os.StrictMode
 import com.google.android.horologist.audio.SystemAudioRepository
 import com.google.android.horologist.media3.audio.AudioOutputSelector
 import com.google.android.horologist.media3.audio.BluetoothSettingsOutputSelector
-import com.google.android.horologist.mediasample.AppConfig
-import com.google.android.horologist.mediasample.util.resetAfter
+import com.google.android.horologist.mediasample.ui.AppConfig
+import com.google.android.horologist.mediasample.ui.util.resetAfter
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
