+++ media-sample/src/main/java/com/google/android/horologist/mediasample/di/PlaybackServiceModule.kt
@@ -51,9 +51,9 @@ import com.google.android.horologist.media3.logging.TransferListener
 import com.google.android.horologist.media3.navigation.IntentBuilder
 import com.google.android.horologist.media3.offload.AudioOffloadManager
 import com.google.android.horologist.media3.rules.PlaybackRules
-import com.google.android.horologist.mediasample.AppConfig
-import com.google.android.horologist.mediasample.complication.DataUpdates
-import com.google.android.horologist.mediasample.media.UampMediaLibrarySessionCallback
+import com.google.android.horologist.mediasample.data.service.complication.DataUpdates
+import com.google.android.horologist.mediasample.data.service.playback.UampMediaLibrarySessionCallback
+import com.google.android.horologist.mediasample.ui.AppConfig
 import com.google.android.horologist.networks.data.RequestType
 import com.google.android.horologist.networks.okhttp.NetworkAwareCallFactory
 import dagger.Module
